requirements.txt
flwr[simulation]
flwr-datasets[vision] 
torch 
torchvision 
matplotlib
streamlit
hydra-core
ray